+++ pennylane_lightning/core/src/simulators/lightning_gpu/observables/ObservablesGPU.cpp
@@ -28,3 +28,6 @@ template class Observables::TensorProdObs<StateVectorCudaManaged<double>>;
 
 template class Observables::Hamiltonian<StateVectorCudaManaged<float>>;
 template class Observables::Hamiltonian<StateVectorCudaManaged<double>>;
+
+template class Observables::SparseHamiltonian<StateVectorCudaManaged<float>>;
+template class Observables::SparseHamiltonian<StateVectorCudaManaged<double>>;